Allow RS256 to sign JWT tokens, including backchannel (#1)
* Allow RS256 to sign JWT tokens, including backchannel

Applications that have a one login integration already in place in production.
With a different signing algorithm to ES256(which is the default in this gem),
cannot easily use this gem.

One login production integrations cannot easily change signing algorithms.
As this will require downtime.

So for services that use RS256 and cannot change their integration easily.
This commit changes the gem to allow the client to configure the signing algorithm.

In essence we need changes in two places. When signing in. We need to use the signing
algorithm that the client specified.

And also in the backchannel functionality. We need to allow both ES256 and RS256 to
decode the token. As we don't have the client preference at that point.